+++ consensus/spos/interface.go
@@ -105,7 +105,7 @@ type WorkerHandler interface {
 	// AddReceivedHeaderHandler adds a new handler function for a received header
 	AddReceivedHeaderHandler(handler func(data.HeaderHandler))
 	// AddReceivedProofHandler adds a new handler function for a received proof
-	AddReceivedProofHandler(handler func(ProofHandler))
+	AddReceivedProofHandler(handler func(consensus.ProofHandler))
 	// RemoveAllReceivedMessagesCalls removes all the functions handlers
 	RemoveAllReceivedMessagesCalls()
 	// ProcessReceivedMessage method redirects the received message to the channel which should handle it
@@ -258,14 +258,4 @@ type RoundThresholdHandler interface {
 	SetThreshold(subroundId int, threshold int)
 	FallbackThreshold(subroundId int) int
 	SetFallbackThreshold(subroundId int, threshold int)
-}
-
-// ProofHandler defines the interface for a proof handler
-type ProofHandler interface {
-	GetPubKeysBitmap() []byte
-	GetAggregatedSignature() []byte
-	GetHeaderHash() []byte
-	GetHeaderEpoch() uint32
-	GetHeaderNonce() uint64
-	GetHeaderShardId() uint32
 }
