+++ testscommon/consensus/sposWorkerMock.go
@@ -7,7 +7,6 @@ import (
 	"github.com/multiversx/mx-chain-core-go/data"
 
 	"github.com/multiversx/mx-chain-go/consensus"
-	"github.com/multiversx/mx-chain-go/consensus/spos"
 	"github.com/multiversx/mx-chain-go/p2p"
 )
 
@@ -18,7 +17,7 @@ type SposWorkerMock struct {
 		receivedMessageCall func(ctx context.Context, cnsDta *consensus.Message) bool,
 	)
 	AddReceivedHeaderHandlerCalled         func(handler func(data.HeaderHandler))
-	AddReceivedProofHandlerCalled          func(handler func(proofHandler spos.ProofHandler))
+	AddReceivedProofHandlerCalled          func(handler func(proofHandler consensus.ProofHandler))
 	RemoveAllReceivedMessagesCallsCalled   func()
 	ProcessReceivedMessageCalled           func(message p2p.MessageP2P) error
 	SendConsensusMessageCalled             func(cnsDta *consensus.Message) bool
@@ -48,7 +47,7 @@ func (sposWorkerMock *SposWorkerMock) AddReceivedHeaderHandler(handler func(data
 	}
 }
 
-func (sposWorkerMock *SposWorkerMock) AddReceivedProofHandler(handler func(proofHandler spos.ProofHandler)) {
+func (sposWorkerMock *SposWorkerMock) AddReceivedProofHandler(handler func(proofHandler consensus.ProofHandler)) {
 	if sposWorkerMock.AddReceivedProofHandlerCalled != nil {
 		sposWorkerMock.AddReceivedProofHandlerCalled(handler)
 	}
